Create API_CortellisLab_DataExtraction_1
Cortellis clinical trials intelligence (CCTI) is a data base from Clarivate that contains information of  worldwide clinical trials (covering drugs, biologics, devices and biomarkers)
....

This script was done to optimize the process of manually going to CCTI for a list of clinical trials and get some required fields. It automatize the process to access to the API, and iterate through the tree till reaching the specific branches were the field of interest is found.