Update API_CortellisClinicalTrialsIntelligence_Data extraction
Cortellis Clinical trials Intelligence (CCTI), is product from Clarivate that provides manually curated information on worldwide clinical trials (covering drugs, biologics, devices and biomarkers), press releases, conference abstracts and reports. 

This script was written to automatize the process of gettingthe API, iterating through the tree and reaching the branches with the fields required, for a large list of Clinical trials.